move each profiler to its own file
